chore(package.json) update dependencies versions
Due to security alerts:
- [vm2 Sandbox Escape vulnerability](https://github.com/xlts-dev/angularjs-angular-hybrid-angular-cli/security/dependabot/96)
- [vm2 vulnerable to Inspect Manipulation](https://github.com/xlts-dev/angularjs-angular-hybrid-angular-cli/security/dependabot/97)
- [Cross site scripting in Angular](https://github.com/xlts-dev/angularjs-angular-hybrid-angular-cli/security/dependabot/78)

The first two alerts are related to the installed version 3.9.17 of vm2,
which is a dependency of the dependency of firebase-tools@11.30.0. Both
problems have been fixed in the release of version 3.9.18 of vm2. By
installing firebase-tools v12.3.0, we install the vm2 dependency version
3.9.19, which fixes the mentioned errors.

The last issue is related to packages associated with Angular v15.2.8
and is related to comment handling, where manipulation leads to cross
site scripting. Upgrading the packages to v15.2.9 resolves this issue.